SRA-law-firm-lookup-tool/requirements.txt
fastapi==0.112.*
uvicorn[standard]==0.30.*
python-dotenv==1.0.*
requests==2.32.*